+++ src/core/json-schema/src/formly-json-schema.service.spec.ts
@@ -2,7 +2,7 @@ import { FormlyJsonschema } from './formly-json-schema.service';
 import { JSONSchema7 } from 'json-schema';
 import { FormlyFieldConfig, FormlyFieldProps, FieldArrayType } from '@ngx-formly/core';
 import { FormControl, FormGroup, FormArray } from '@angular/forms';
-import { createComponent, FormlyInputModule } from '@ngx-formly/core/testing';
+import { createComponent, FormlyInputModule, ɵCustomEvent } from '@ngx-formly/core/testing';
 import { Component } from '@angular/core';
 
 const renderComponent = ({ schema, model }: { schema: JSONSchema7; model?: any }) => {
@@ -366,7 +366,7 @@ describe('Service: FormlyJsonschema', () => {
         const config = formlyJsonschema.toFieldConfig(numSchema);
         expect(config.props.minItems).toBe(numSchema.minItems);
 
-        const minItemsValidator = (model: any) => config.validators.minItems(new FormControl(), { model });
+        const minItemsValidator = (model: any) => config.validators.minItems(new FormControl(model), { model });
         expect(minItemsValidator(undefined)).toBeTrue();
         expect(minItemsValidator([1])).toBeFalse();
         expect(minItemsValidator([])).toBeFalse();
@@ -405,7 +405,7 @@ describe('Service: FormlyJsonschema', () => {
         const config = formlyJsonschema.toFieldConfig(numSchema);
         expect(config.props.maxItems).toBe(numSchema.maxItems);
 
-        const maxItemsValidator = (model: any) => config.validators.maxItems(new FormControl(), { model });
+        const maxItemsValidator = (model: any) => config.validators.maxItems(new FormControl(model), { model });
         expect(maxItemsValidator(undefined)).toBeTrue();
         expect(maxItemsValidator([1, 2, 3])).toBeFalse();
         expect(maxItemsValidator([1, 2])).toBeTrue();
@@ -420,7 +420,7 @@ describe('Service: FormlyJsonschema', () => {
         const config = formlyJsonschema.toFieldConfig(numSchema);
         expect(config.props.uniqueItems).toBeTrue();
 
-        const uniqueItemsValidator = (model: any) => config.validators.uniqueItems(new FormControl(), { model });
+        const uniqueItemsValidator = (model: any) => config.validators.uniqueItems(new FormControl(model), { model });
         expect(uniqueItemsValidator(undefined)).toBeTrue();
         expect(uniqueItemsValidator([1, 2, 3])).toBeTrue();
         expect(uniqueItemsValidator([1, 2, 2])).toBeFalse();
@@ -980,9 +980,9 @@ describe('Service: FormlyJsonschema', () => {
 
         const expectedConfig: FormlyFieldConfig = {
           type: 'array',
-          defaultValue: undefined,
-          props: { ...emmptyFieldProps, required: true },
-          templateOptions: { ...emmptyFieldProps, required: true },
+          defaultValue: [],
+          props: { ...emmptyFieldProps },
+          templateOptions: { ...emmptyFieldProps },
           fieldArray: expect.any(Function),
           validators: expectTypeValidator(['array']),
         };
@@ -1997,6 +1997,47 @@ describe('Service: FormlyJsonschema', () => {
       });
 
       expect(field).toBeDefined();
+    });
+  });
+
+  describe('FormlyJsonSchemaOptions map data', () => {
+    it('should set undefined when number type input is empty', () => {
+      const { field, query } = renderComponent({
+        schema: { type: 'integer' },
+      });
+
+      query('input').triggerEventHandler('input', ɵCustomEvent({ value: 'eeee' }));
+      expect(field.formControl.value).toEqual('eeee');
+
+      query('input').triggerEventHandler('input', ɵCustomEvent({ value: '' }));
+      expect(field.formControl.value).toEqual(undefined);
+
+      query('input').triggerEventHandler('input', ɵCustomEvent({ value: '2e3' }));
+      expect(field.formControl.value).toEqual(2000);
+    });
+
+    it('should set non required string to undefined when is empty', () => {
+      const { field } = renderComponent({
+        schema: { type: 'string' },
+      });
+
+      const parser = field.parsers[0] as any;
+
+      expect(parser('', field)).toEqual(undefined);
+      field.props.required = true;
+      expect(parser('', field)).toEqual('');
+    });
+
+    it('should set default value for array type items', () => {
+      const { field } = renderComponent({
+        model: [undefined],
+        schema: {
+          type: 'array',
+          items: { type: 'object' },
+        },
+      });
+
+      expect(field.fieldGroup[0].defaultValue).toEqual({});
     });
   });
 });